vala-mode: support multiline strings
Unfortunately, it seems to be tricky to add support for Vala’s multiline
string syntax to CC Mode, because that offers only either all strings being
multiline, or strings preceded by a certain character. Of course, Vala is in
neither category.

For now, define `c-multiline-string-start-char` to `t`, which means that all
strings are allowed to be multiline. This means that an
incorrectly-terminated normal string will be fontified as a long string, not
as an error, but nonetheless it should be obvious to the user what has gone
wrong; meanwhile, long strings are correctly fontified.